// src/types/form.types.ts
export type ContactFormData = {
  name: string;
  surname: string;
  email: string;
  phone: string;
  message: string;
  kvkk: boolean;
};

export type QuoteFormData = {
  name: string;
  surname: string;
  email: string;
  phone: string;
  company: string;
  taxNumber: string;
  product: string;
  message: string;
  kvkk: boolean;
  consent: boolean;
};